Revert "moves logger to y"
This reverts commit 0b81618b76cc7c32b17bfd6d54d9c7141ec60348.
